\name{EnsDb.Gmax.v62}
\docType{package}

\alias{EnsDb.Gmax.v62-package}
\alias{EnsDb.Gmax.v62}
\alias{EnsDb.Gmax.v62}


\title{Ensembl based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at Ensembl on
  Thu May 21 11:06:16 2026 and based on the 62
}

\author{Kabilan S}



\examples{
## load the library
##library(EnsDb.Gmax.v62)
## list the contents that are loaded into memory
ls('package:EnsDb.Gmax.v62')
## show the db object that is loaded by calling it's name
EnsDb.Gmax.v62

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
